visionandbrowseragents/requirements.txt
Pillow
requests
smolagents
smolagents[litellm]

smolagents[all]
helium
selenium
python-dotenv